+++ 1.11.2/src/main/java/net/geforcemods/securitycraft/items/ItemScannerDoor.java
@@ -44,7 +44,7 @@ public class ItemScannerDoor extends Item
 
             if (playerIn.canPlayerEdit(pos, facing, stack) && mod_SecurityCraft.scannerDoor.canPlaceBlockAt(worldIn, pos))
             {
-                EnumFacing enumfacing = EnumFacing.fromAngle((double)playerIn.rotationYaw);
+                EnumFacing enumfacing = EnumFacing.fromAngle(playerIn.rotationYaw);
                 int i = enumfacing.getFrontOffsetX();
                 int j = enumfacing.getFrontOffsetZ();
                 boolean flag = i < 0 && hitZ < 0.5F || i > 0 && hitZ > 0.5F || j < 0 && hitX > 0.5F || j > 0 && hitX < 0.5F;
